+++ sdk/appcontainers/azure-mgmt-appcontainers/generated_samples/managed_certificate_delete.py
@@ -29,14 +29,13 @@ def main():
         subscription_id="34adfa4f-cedf-4dc0-ba29-b6d1a69ab345",
     )
 
-    response = client.managed_certificates.delete(
+    client.managed_certificates.delete(
         resource_group_name="examplerg",
         environment_name="testcontainerenv",
         managed_certificate_name="certificate-firendly-name",
     )
-    print(response)
 
 
-# x-ms-original-file: specification/app/resource-manager/Microsoft.App/preview/2022-11-01-preview/examples/ManagedCertificate_Delete.json
+# x-ms-original-file: specification/app/resource-manager/Microsoft.App/stable/2023-05-01/examples/ManagedCertificate_Delete.json
 if __name__ == "__main__":
     main()