+++ sdk/ml/azure-ai-ml/azure/ai/ml/_schema/monitoring/monitor_definition.py
@@ -8,6 +8,7 @@ from marshmallow import fields, post_load
 
 from azure.ai.ml.constants._monitoring import AZMONITORING
 from azure.ai.ml._schema.monitoring.target import MonitoringTargetSchema
+from azure.ai.ml._schema.monitoring.compute import ServerlessSparkComputeSchema
 from azure.ai.ml._schema.monitoring.signals import (
     DataDriftSignalSchema,
     DataQualitySignalSchema,
@@ -18,11 +19,10 @@ from azure.ai.ml._schema.monitoring.signals import (
 from azure.ai.ml._schema.monitoring.alert_notification import AlertNotificationSchema
 from azure.ai.ml._schema.core.fields import NestedField, UnionField, StringTransformedEnum
 from azure.ai.ml._schema.core.schema import PatchedSchemaMeta
-from azure.ai.ml._schema.spark_resource_configuration import SparkResourceConfigurationSchema
 
 
 class MonitorDefinitionSchema(metaclass=PatchedSchemaMeta):
-    compute = NestedField(SparkResourceConfigurationSchema)
+    compute = NestedField(ServerlessSparkComputeSchema)
     monitoring_target = NestedField(MonitoringTargetSchema)
     monitoring_signals = fields.Dict(
         keys=fields.Str(),