+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/__init__.py
@@ -50,7 +50,7 @@ from ._credentials import (
     UserIdentityConfiguration,
     UsernamePasswordConfiguration,
     AccessKeyConfiguration,
-    ApiKeyConfiguration
+    ApiKeyConfiguration,
 )
 from ._datastore.adls_gen1 import AzureDataLakeGen1Datastore
 from ._datastore.azure_storage import AzureBlobDatastore, AzureDataLakeGen2Datastore, AzureFileDatastore
@@ -174,7 +174,6 @@ from ._monitoring.signals import (
     PredictionDriftSignal,
     FeatureAttributionDriftSignal,
     CustomMonitoringSignal,
-    TargetDataset,
     MonitorFeatureFilter,
     DataSegment,
 )
@@ -407,7 +406,6 @@ __all__ = [
     "PredictionDriftSignal",
     "FeatureAttributionDriftSignal",
     "CustomMonitoringSignal",
-    "TargetDataset",
     "MonitorFeatureFilter",
     "DataSegment",
     "MonitoringTarget",