+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_builders/condition_node.py
@@ -52,7 +52,13 @@ class ConditionNode(ControlFlowNode):
 
     @classmethod
     def _create_instance_from_schema_dict(cls, loaded_data: Dict) -> "ConditionNode":
-        """Create a condition node instance from schema parsed dict."""
+        """Create a condition node instance from schema parsed dict.
+
+        :param loaded_data: The loaded data
+        :type loaded_data: Dict
+        :return: The ConditionNode node
+        :rtype: ConditionNode
+        """
         return cls(**loaded_data)
 
     @property