+++ sdk/storage/azure-mgmt-storage/azure/mgmt/storage/v2018_07_01/operations/__init__.py
@@ -12,7 +12,6 @@ from ._storage_accounts_operations import StorageAccountsOperations
 from ._usages_operations import UsagesOperations
 from ._blob_services_operations import BlobServicesOperations
 from ._blob_containers_operations import BlobContainersOperations
-from ._management_policies_operations import ManagementPoliciesOperations
 
 from ._patch import __all__ as _patch_all
 from ._patch import *  # pylint: disable=unused-wildcard-import
@@ -25,7 +24,6 @@ __all__ = [
     "UsagesOperations",
     "BlobServicesOperations",
     "BlobContainersOperations",
-    "ManagementPoliciesOperations",
 ]
 __all__.extend([p for p in _patch_all if p not in __all__])
 _patch_sdk()